Split up the modals
